requirements.txt
appdirs==1.4.4
astroid==2.4.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
google-api-core==1.26.1
google-auth==1.28.0
google-cloud-bigquery==2.13.1
google-cloud-core==1.6.0
google-cloud-datastore==2.1.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
libcst==0.3.17
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
virtualenv==20.4.3
Werkzeug==1.0.1
wrapt==1.12.1
